% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa_category.R
\name{gpa_category}
\alias{gpa_category}
\title{Categorize a student's GPA}
\usage{
gpa_category(gpa)
}
\arguments{
\item{gpa}{A numeric GPA value, usually between 0.0 and 4.0.}
}
\value{
A character value describing the GPA category.
}
\description{
This function takes a GPA value and returns a simple category.
}
\examples{
gpa_category(3.8)
gpa_category(2.4)

}
