% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_category.R
\name{grade_category}
\alias{grade_category}
\title{Convert numeric grades to letter categories}
\usage{
grade_category(grade)
}
\arguments{
\item{grade}{A numeric value representing a student's grade}
}
\value{
A character value representing the letter grade
}
\description{
This function takes a numeric grade (0-100) and returns
a letter grade category (A, B, C, D, F).
}
\examples{
grade_category(95)
grade_category(72)

}
