% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_report.R
\name{missing_report}
\alias{missing_report}
\title{Report missing values in a dataset}
\usage{
missing_report(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame showing missing values per column
}
\description{
This function takes a data frame and returns the number
of missing (NA) values in each column.
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(NA, NA, 2))
missing_report(df)

}
