% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_profile.R
\name{print.student_profile}
\alias{print.student_profile}
\title{Print a student profile}
\usage{
\method{print}{student_profile}(x, ...)
}
\arguments{
\item{x}{A student_profile object.}

\item{...}{Additional arguments.}
}
\value{
No return value, called for printing.
}
\description{
This function prints a student_profile object in a clean format.
}
\examples{
profile <- student_profile("Joe", 3.5, 88)
print(profile)

}
