% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_profile.R
\name{student_profile}
\alias{student_profile}
\title{Create a student profile}
\usage{
student_profile(name, gpa, grade)
}
\arguments{
\item{name}{A character value for the student's name.}

\item{gpa}{A numeric GPA value.}

\item{grade}{A numeric grade value.}
}
\value{
An object of class student_profile.
}
\description{
This function creates a simple student profile object using an S3 class.
}
\examples{
student_profile("Joe", 3.5, 88)

}
